manifests: use raspberry pi sysint repo / remove obsolete copy
The repository mentioned in the manifest was not made public
as part of the open sourcing process, as it's specific files were moved
inside the layer.

However, we still need some test and diagnostic related files
from the Raspberry Pi repo, and due to externalsrc,
bitbake ignores the SRC_URI for code.rdkcentral.com in
our sysint-broadband.bbapend.